+++ jax/_src/deprecations.py
@@ -128,3 +128,4 @@ def warn(deprecation_id: str, message: str, stacklevel: int, *,
 register('jax-array-numpy-dtype')
 register('jax-nn-one-hot-float-input')
 register('jax-numpy-astype-complex-to-real')
+register('jax-array-positional-args')